Summary:
- the stride logic is improved by using displacements instead of velocities. Reference position and sweepLength are added to bounds.
- the sweepLength of each body is updated in collider::action() according to collider's targetInterv, so that ideally all bodies leave their bounds at the same iteration (more or less the same idea as in the old velocityBin but with per-body logic)
- a number of optimizations in the collider and interaction dispatcher. Namely, the handling of virtual interactions is a lot faster, allowing a larger number of them for the optimal verletDist.
- timings reported at https://yade-dem.org/wiki/Colliders_performace (please add your reports there), roughly x3 speedup vs. bzr2999 in triaxial simulations
- A new ZECollider is introduced, using CGAL's dD intersection algorithm. It is relatively slow and therefore not recommended, but the code is very short and it may be worth a try when parallel CGAL is released.
- Added references to some CGAL packages and a few fixes in references.

+++ pkg/dem/TesselationWrapper.cpp
@@ -158,7 +158,7 @@ void TesselationWrapper::insertSceneSpheres(bool reset)
 // 	clock.top("Triangulation");
 }
 
-double TesselationWrapper::Volume(unsigned int id) {return (Tes->Max_id() > id) ? Tes->Volume(id) : -1;}
+double TesselationWrapper::Volume(unsigned int id) {return ((unsigned int) Tes->Max_id() > id) ? Tes->Volume(id) : -1;}
 
 bool TesselationWrapper::insert(double x, double y, double z, double rad, unsigned int id)
 {
